feature/VIH-6620 : fixing sonar cloud issues
