--- ServiceWebsite/ServiceWebsite/ClientApp/src/app/pages/checking-video-hearing/checking-video-hearing.component.html
@@ -1,26 +1,0 @@
-<app-crest-blue-panel>
-    <h1 class="govuk-heading-xl text-white" i18n="@@aboutHearing_h1_1">Checking a video hearing will work on your computer</h1>
-    <p class="govuk-body govuk-!-margin-bottom-5 text-white" i18n="@@aboutHearing_p_1">
-        We'll now take you through a short check to see if your computer is suitable for a video hearing.
-    </p>
-    <p class="govuk-body govuk-!-margin-bottom-5 text-white" i18n="@@aboutHearing_p_2">
-        You'll need a laptop or desktop computer with a camera and microphone, or an Apple iPad. You cannot use a mobile phone or other
-        tablet computer.
-    </p>
-    <p class="govuk-body govuk-!-margin-bottom-5 text-white" i18n="@@aboutHearing_p_3">
-        Please run this equipment check on the computer or iPad you'll use for your hearing.
-    </p>
-    <p class="govuk-body govuk-!-margin-bottom-5 text-white">
-        <a class="govuk_link text-white"
-           i18n="@@aboutHearing_a_1"
-           href="https://www.gov.uk/guidance/hmcts-video-hearings-service-guidance-for-joining-a-hearing"
-           rel="noopener noreferrer"
-           target="_blank">
-            More information about your computer and internet connection
-        </a>
-    </p>
-    <button type="button" i18n="@@aboutHearing_btn_next" class="govuk-button govuk-button--secondary" (click)="continue()" id="continue">
-        Next
-    </button>
-</app-crest-blue-panel>
-<app-contact-us></app-contact-us>
